disable db on build
