+++ .github/workflows/build.yaml
@@ -10,16 +10,6 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
-    services:
-      ehrbase-db:
-        image: ehrbase/ehrbase-postgres:13.4
-        ports:
-          - 5432:5432
-        env:
-          POSTGRES_USER: postgres
-          POSTGRES_PASSWORD: postgres
-          EHRBASE_USER: ehrbase
-          EHRBASE_PASSWORD: ehrbase
     steps:
       - name: Checkout
         uses: actions/checkout@v2
