+++ src/main/java/org/ehrbase/webtester/config/jooq/PersistenceConfig.java
@@ -10,6 +10,7 @@ import org.jooq.impl.DefaultDSLContext;
 import org.jooq.impl.DefaultExecuteListener;
 import org.jooq.impl.DefaultExecuteListenerProvider;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
@@ -21,6 +22,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @Configuration
 @EnableTransactionManagement
+@ConditionalOnProperty(prefix = "loader", name = "enabled", havingValue = "true")
 public class PersistenceConfig {
 
   static class ExceptionTranslator extends DefaultExecuteListener {
